+++ core/src/main/scala/net/ruippeixotog/scalascraper/dsl/DSL.scala
@@ -5,15 +5,15 @@ import net.ruippeixotog.scalascraper.scraper._
 
 object DSL extends ImplicitConversions with ScrapingOps {
 
-  def extractor(cssQuery: String): HtmlExtractor[Element, Iterable[String]] = SimpleExtractor(cssQuery)
+  def extractor(cssQuery: String): HtmlExtractor[Element, Iterable[String]] = HtmlExtractor(cssQuery)
 
   def extractor[E <: Element, C](cssQuery: String, contentExtractor: ElementQuery[E] => C): HtmlExtractor[E, C] =
-    SimpleExtractor(cssQuery, contentExtractor)
+    HtmlExtractor(cssQuery, contentExtractor)
 
   def extractor[E <: Element, C, A](cssQuery: String, contentExtractor: ElementQuery[E] => C, contentParser: C => A): HtmlExtractor[E, A] =
-    SimpleExtractor(cssQuery, contentExtractor, contentParser)
+    HtmlExtractor(cssQuery, contentExtractor, contentParser)
 
-  val validator = SimpleValidator
+  val validator = HtmlValidator
 
   val Extract = ContentExtractors
   val Parse = ContentParsers