+++ core/src/main/scala/net/ruippeixotog/scalascraper/dsl/ImplicitConversions.scala
@@ -1,16 +1,16 @@
 package net.ruippeixotog.scalascraper.dsl
 
 import net.ruippeixotog.scalascraper.model._
-import net.ruippeixotog.scalascraper.scraper.{ HtmlExtractor, SimpleExtractor }
+import net.ruippeixotog.scalascraper.scraper.HtmlExtractor
 import net.ruippeixotog.scalascraper.util._
 
 trait ImplicitConversions {
 
   implicit def cssQueryAsExtractor(cssQuery: String): HtmlExtractor[Element, Iterable[String]] =
-    SimpleExtractor(cssQuery)
+    HtmlExtractor(cssQuery)
 
   implicit def contentExtractorAsExtractor[E <: Element, C](contentExtractor: ElementQuery[E] => C): String => HtmlExtractor[E, C] = {
-    cssQuery: String => SimpleExtractor(cssQuery, contentExtractor)
+    cssQuery: String => HtmlExtractor(cssQuery, contentExtractor)
   }
 
   implicit def elementAsElementQuery[E <: Element.Upper[E]](elem: E) = ElementQuery(elem)