+++ modules/config/src/main/scala/net/ruippeixotog/scalascraper/config/ConfigHtmlValidator.scala
@@ -1,6 +1,6 @@
 package net.ruippeixotog.scalascraper.config
 
-import scala.util.Try
+import scala.util.{ Failure, Success, Try }
 
 import com.typesafe.config.Config
 
@@ -11,21 +11,27 @@ import net.ruippeixotog.scalascraper.scraper._
 object ConfigHtmlValidator {
 
   def apply[R](conf: Config)(implicit reader: ConfigReader[R]): HtmlValidator[Element, R] = {
+    val result = Try(reader.read(conf, "status"))
+
     if (conf.hasPath("exists")) {
-      val extractor = SimpleExtractor(
+      val extractor = HtmlExtractor(
         conf.getString("select.query"),
         ContentExtractors.elements, ContentParsers.asIs[ElementQuery[Element]])
 
       val matcher = conf.getBoolean("exists") == (_: Iterable[Element]).nonEmpty
-      val result = Try(reader.read(conf, "status")).toOption
-      SimpleValidator(extractor, matcher, result)
+      result match {
+        case Success(res) => HtmlValidator(extractor, res)(matcher)
+        case Failure(_) => HtmlValidator(extractor)(matcher)
+      }
 
     } else {
       val extractor = ConfigHtmlExtractor[String](conf.getConfig("select"))
       val matchIfTrue = if (conf.hasPath("invert")) !conf.getBoolean("invert") else true
       val matcher = conf.getString("match").r.pattern.matcher(_: String).matches() == matchIfTrue
-      val result = Try(reader.read(conf, "status")).toOption
-      SimpleValidator(extractor, matcher, result)
+      result match {
+        case Success(res) => HtmlValidator(extractor, res)(matcher)
+        case Failure(_) => HtmlValidator(extractor)(matcher)
+      }
     }
   }
 }