+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 - Breaking changes
   - `HtmlExtractor`, `HtmlValidator` and `ElementQuery` now have an additional type parameter for the type of `Element`
     they work on. Filling it with `Element` (which is a superclass of all elements) should be enough for them to work
-    with all 1.x code;
+    with all source code using scala-scraper 1.x;
+  - `SimpleExtractor` and `SimpleValidator` are now deprecated. The classes remain available for the time being, but DSL
+    methods that returned those classes now return only `HtmlExtractor` and `HtmlValidator` instances;
   - Methods for loading extractors and validators from a config were extracted to a seaparate module. In order to use
     them users must add `scala-scraper-config` to their SBT dependencies and import
     `net.ruippeixotog.scalascraper.config.dsl.DSL._`.